spliting out classes glyphScene, glyphView, handleLineItem, offCurvePointItem and onCurvePointItem to their own files
